+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # @xellar/sdk
 
+## 4.6.1
+
+### Patch Changes
+
+- [`ed7592d`](https://github.com/Xellar-Protocol/ew-sdk/commit/ed7592da65dc04a9d2ef8a62d540f1105e1c9db6) Thanks [@iyansr](https://github.com/iyansr)! - Make Only Chain ID Required
+
+## 4.6.0
+
+### Minor Changes
+
+- [`d5f10ef`](https://github.com/Xellar-Protocol/ew-sdk/commit/d5f10ef0f6a79111f11185f4dc133df6a6ec795e) Thanks [@iyansr](https://github.com/iyansr)! - Add Sign Authorization
+
+- [#58](https://github.com/Xellar-Protocol/ew-sdk/pull/58) [`0d36c30`](https://github.com/Xellar-Protocol/ew-sdk/commit/0d36c309023ff17295f0f1c13e5f6cadb9a45115) Thanks [@yoms07](https://github.com/yoms07)! - add get user op by id
+
 ## 4.5.2
 
 ### Patch Changes